Use user id or email depending on if user exists
Previously we had made a change to rely on the user id for avalara's
customer code value. However, with the adding guest checkout we may
not have a user id for all customers. So this change will follow the
standard used in the certified avatax gem that uses the id when we have
a user and uses the email when we do not. I am also adding a limit to
make sure this value doesn't exceed 50 characters since that is the max
allowed value by avalara. This should allow us to calculate tax's for
all of orders including guest orders.