Allow for guest orders without emails yet for taxes
With the addition of apple pay and stripe payment request button we
would like for our customers to have their taxes calculated before we
add an email to the user. This should allow us to do that by creating a
last option for when the user or email does not exist on the order and
we need a customer code to calculate taxes.