Reduce the number of calls to Spree::ItemAdjustments#update
With this change we'll reduce the number of times we attempt
to update the adjustments associated with a line item, this
operation is pretty expensive as it requires us to re-evaluate
every adjustment associated with the line item.

Since we eventually call Spree::Order#update!, line items
will eventually update all their associated adjustments
and end up with the right numbers.